requirements.txt
Django~=2.0.7
python-binance~=0.6.9
pyyaml~=3.13
django-celery-beat~=1.1.1
django-background-tasks~=1.1.13
django-periodically~=0.3.0